Complete SpaceFOM initialization and execution sequence (final part)

+++ src/test/java/io/github/vega/outpost/components/PositionComponent.java
@@ -4,7 +4,7 @@ import com.badlogic.ashley.core.Component;
 
 public class PositionComponent implements Component
 {
-	public int x = 0;
-	public int y = 0;
-	public int z = 0;
+	public double x = 0;
+	public double y = 0;
+	public double z = 0;
 }
